Update archive-manager testing approach
